AB#27 create control panel
